Extract regex literal set parsing
